@@ -381,12 +381,21 @@ def extract_particles(
 
     if relion5_compat:
         dims = np.array(job.tomo_shape)
-        center = dims / 2  # we approximate the center, the correct way is to use
-        # relion5's unbinned dimensions and calculate: center = (ubinned_dims / 2) - 1
+        if "relion5_binning" in job.metadata:
+            binning = job.metadata["relion5_binning"]
+            center = (dims * binning) / 2 - 1
+            voxel_size = job.metadata["relion5_ts_ps"]
+        else:
+            # we approximate the center, the correct way is to use
+            # relion5's unbinned dimensions and calculate:
+            # center = (ubinned_dims / 2) - 1 as above
+            binning = 1
+            center = dims / 2
+            voxel_size = job.voxel_size
         output["rlnCoordinateX"], output["rlnCoordinateY"], output["rlnCoordinateZ"] = (
-            (output["rlnCoordinateX"] - center[0]) * job.voxel_size,
-            (output["rlnCoordinateY"] - center[1]) * job.voxel_size,
-            (output["rlnCoordinateZ"] - center[2]) * job.voxel_size,
+            (output["rlnCoordinateX"] * binning - center[0]) * voxel_size,
+            (output["rlnCoordinateY"] * binning - center[1]) * voxel_size,
+            (output["rlnCoordinateZ"] * binning - center[2]) * voxel_size,
         )
         column_change = {
             "rlnCoordinateX": "rlnCenteredCoordinateXAngst",
