@@ -1,4 +1,5 @@
 import unittest
+import sys
 import pathlib
 import numpy as np
 import cupy as cp
@@ -6,8 +7,10 @@ import logging
 from shutil import which
 from contextlib import redirect_stdout, redirect_stderr
 from io import StringIO
+from tempfile import TemporaryDirectory
 from pytom_tm import entry_points
 from pytom_tm import io
+from testing_utils import chdir
 
 # (command line function, function in entry_points file)
 ENTRY_POINTS_TO_TEST = [
@@ -30,7 +33,6 @@ TEST_DATA = pathlib.Path(__file__).parent.joinpath("test_data")
 TEMPLATE = TEST_DATA.joinpath("template.mrc")
 MASK = TEST_DATA.joinpath("mask.mrc")
 TOMOGRAM = TEST_DATA.joinpath("tomogram.mrc")
-DESTINATION = TEST_DATA.joinpath("output")
 TILT_ANGLES = TEST_DATA.joinpath("angles.rawtlt")
 TILT_ANGLES_MULTI_COLUMN = TEST_DATA.joinpath("angles_multi_column.rawtlt")
 DOSE = TEST_DATA.joinpath("test_dose.txt")
@@ -38,6 +40,10 @@ DEFOCUS = TEST_DATA.joinpath("defocus.txt")
 DEFOCUS_IMOD = (
     pathlib.Path(__file__).parent.joinpath("Data").joinpath("test_imod.defocus")
 )
+WARP_XML = pathlib.Path(__file__).parent.joinpath(
+    "Data/warptools_xml_example/gs04_ts_003.xml"
+)
+
 RELION5_TOMOGRAMS_STAR = pathlib.Path(__file__).parent.joinpath(
     "Data/relion5_project_example/Tomograms/job009/tomograms.star"
 )
@@ -56,10 +62,21 @@ def prep_argv(arg_dict):
     return argv
 
 
+class TestParseArgv(unittest.TestCase):
+    def test_parse_argv(self):
+        out = entry_points._parse_argv()
+        # test behavior by repeating the behavior
+        self.assertEqual(sys.argv[1:], out)
+        inp = ["test1", "test2"]
+        out = entry_points._parse_argv(inp)
+        for i in inp:
+            self.assertIn(i, out)
+
+
 class TestEntryPoints(unittest.TestCase):
     @classmethod
     def setUpClass(cls) -> None:
-        DESTINATION.mkdir(parents=True)
+        TEST_DATA.mkdir(parents=True)
         io.write_mrc(TEMPLATE, np.zeros((5, 5, 5), dtype=np.float32), 1)
         io.write_mrc(MASK, np.zeros((5, 5, 5), dtype=np.float32), 1)
         io.write_mrc(TOMOGRAM, np.zeros((10, 10, 10), dtype=np.float32), 1)
@@ -83,10 +100,13 @@ class TestEntryPoints(unittest.TestCase):
         TILT_ANGLES_MULTI_COLUMN.unlink()
         DOSE.unlink()
         DEFOCUS.unlink()
-        for f in DESTINATION.iterdir():
-            f.unlink()  # should test specific output?
-        DESTINATION.rmdir()
         TEST_DATA.rmdir()
+
+    def setUp(self):
+        # set up a destination temp dir
+        tempdir = TemporaryDirectory()
+        self.outputdir = pathlib.Path(tempdir.name)
+        self.addCleanup(tempdir.cleanup)
 
     def test_entry_points_exist(self):
         for cli, fname in ENTRY_POINTS_TO_TEST:
@@ -102,12 +122,105 @@ class TestEntryPoints(unittest.TestCase):
             # check if the system return code is 0 (success)
             self.assertEqual(ex.exception.code, 0)
 
+    def test_create_mask(self):
+        defaults = {
+            "-b": "60",
+            "-r": "12",
+        }
+        default_output_name = f"mask_b{defaults['-b']}px_r{float(defaults['-r'])}px.mrc"
+
+        def start(arg_dict):
+            entry_points.pytom_create_mask(prep_argv(arg_dict))
+
+        # Test defaults, do change to temp dir
+        with chdir(self.outputdir):
+            start(defaults)
+            # Make sure default file exists
+            self.assertTrue(pathlib.Path(default_output_name).exists())
+            # Make sure it has the expected size
+            meta_data = io.read_mrc_meta_data(default_output_name)
+            self.assertEqual(len(meta_data["shape"]), 3)
+            for n in meta_data["shape"]:
+                self.assertEqual(n, 60)
+
+        # Check we fail loud if only one of the two options is given
+        for i in ["--radius-minor1", "--radius-minor2"]:
+            inp = defaults.copy()
+            inp[i] = "6"
+            with self.assertRaisesRegex(ValueError, f"Only got {i}"):
+                start(inp)
+
+        # smoke test eliptical mask and uneven box
+        inp = defaults.copy()
+        inp["--radius-minor1"] = "6"
+        inp["--radius-minor2"] = "8"
+        inp["-b"] = "55"
+        inp["-o"] = str(self.outputdir / "mask_ellipse.mrc")
+        start(inp)
+
+        self.assertTrue(self.outputdir.joinpath("mask_ellipse.mrc").exists())
+        meta_data = io.read_mrc_meta_data(self.outputdir.joinpath("mask_ellipse.mrc"))
+        self.assertEqual(len(meta_data["shape"]), 3)
+        for n in meta_data["shape"]:
+            self.assertEqual(n, 55)
+
+    def test_create_template(self):
+        defaults = {"-i": str(TEMPLATE), "--output-voxel-size-angstrom": "1.000"}
+        default_output_name = f"template_{TEMPLATE.stem}_{float(1)}A.mrc"
+
+        def start(arg_dict):
+            entry_points.pytom_create_template(prep_argv(arg_dict))
+
+        # Test defaults, do change to temp dir
+        with chdir(self.outputdir):
+            start(defaults)
+            # Make sure default file exists
+            self.assertTrue(pathlib.Path(default_output_name).exists())
+            # Make sure it has the expected size
+            meta_data = io.read_mrc_meta_data(default_output_name)
+            self.assertEqual(len(meta_data["shape"]), 3)
+            for n in meta_data["shape"]:
+                self.assertEqual(n, 5)
+            self.assertEqual(meta_data["voxel_size"], 1.0)
+
+        # Test warning and no warning on input voxelsize
+        args = defaults.copy()
+        output = self.outputdir / "rounded_template.mrc"
+        args["-o"] = f"{output}"
+        # should round correctly at 3 digits
+        args["--input-voxel-size"] = "0.9999"
+        args["--log-test"] = ""
+        with self.assertNoLogs(level=logging.WARNING):
+            start(args)
+        self.assertTrue(output.exists())
+
+        # Now test with the warning
+        args = defaults.copy()
+        output = self.outputdir / "wrong_rounded_template.mrc"
+        args["-o"] = f"{output}"
+        # use 6 as 5 still rounds corrctly due to python bankers' round
+        args["--input-voxel-size"] = "1.0006"
+        # Don't try to invent pixels
+        args["--output-voxel-size-angstrom"] = "2.0"
+        args["--log-test"] = ""
+        with self.assertLogs(level="WARNING") as cm:
+            start(args)
+        self.assertEqual(len(cm.output), 1)
+        self.assertIn("voxel size does not match", cm.output[0])
+        self.assertTrue(output.exists())
+
+        # Test failure on trying to get more resolution
+        args = defaults.copy()
+        args["--output-voxel-size-angstrom"] = "0.5"
+        with self.assertRaisesRegex(NotImplementedError, "smaller voxel size"):
+            start(args)
+
     def test_match_template(self):
         defaults = {
             "-t": str(TEMPLATE),
             "-m": str(MASK),
             "-v": str(TOMOGRAM),
-            "-d": str(DESTINATION),
+            "-d": str(self.outputdir),
             "--angular-search": "35",
             "--tilt-angles": str(TILT_ANGLES),
             "--per-tilt-weighting": "",
@@ -185,11 +298,11 @@ class TestEntryPoints(unittest.TestCase):
         start(arguments)
         # these files will only exist if the test managed to set the logging correctly
         self.assertTrue(
-            DESTINATION.joinpath("template_psf.mrc").exists(),
+            self.outputdir.joinpath("template_psf.mrc").exists(),
             msg="File should exist in debug mode",
         )
         self.assertTrue(
-            DESTINATION.joinpath("template_convolved.mrc").exists(),
+            self.outputdir.joinpath("template_convolved.mrc").exists(),
             msg="File should exist in debug mode",
         )
 
@@ -217,6 +330,24 @@ class TestEntryPoints(unittest.TestCase):
         arguments["-g"] = "0 0 0"
         with self.assertRaisesRegex(ValueError, r"4 tomogram pieces.*3 GPUs"):
             start(arguments)
+
+        # test warp xml metadata reading
+        arguments = defaults.copy()
+        [
+            arguments.pop(x)
+            for x in [
+                "--tilt-angles",
+                "--per-tilt-weighting",
+                "--dose-accumulation",
+                "--defocus",
+                "--amplitude-contrast",
+                "--spherical-aberration",
+                "--voltage",
+            ]
+        ]
+        arguments["-v"] = str(RELION5_TOMOGRAM)
+        arguments["--warp-xml-file"] = str(WARP_XML)
+        start(arguments)
 
         # test relion5 metadata reading
         arguments = defaults.copy()
